Created FollowController and migrated related functions from UserPublicInfoController to there
